+++ pysui_gadgets/__init__.py
@@ -14,14 +14,14 @@
 """Gadget package initializer."""
 
 import sys
-from pysui.version import __version__ as base_ver
+import pysui
 from pysui_gadgets.version import __version__ as gadget_ver
 
-pysui_ver = [int(x) for x in base_ver.split(".")]
+pysui_ver = [int(x) for x in pysui.version.__version__.split(".")]
 pysui_gadget_ver = [int(x) for x in gadget_ver.split(".")]
 
 if sys.version_info < (3, 10):
     raise EnvironmentError("Python 3.10 or above is required")
 
 if pysui_ver[1] < 4:
-    raise EnvironmentError(f"pysui 0.4.0 or above is required, found {base_ver}")
+    raise EnvironmentError(f"pysui 0.4.0 or above is required, found {pysui.version.__version__}")