+++ pysui_gadgets/dsl/dslgen.py
@@ -22,7 +22,7 @@ from pysui_gadgets.dsl.ir.modir import IRBuilder
 from pysui_gadgets.dsl.generators.package_gen import PackageGen
 
 _TEMPLATE_PATH: Path = Path(os.path.abspath(__file__)).parent.joinpath("templates")
-_MODEL_PATH = _TEMPLATE_PATH.joinpath("sync_model.py")
+# _MODEL_PATH = _TEMPLATE_PATH.joinpath("sync_model.py")
 
 
 def main():
@@ -40,7 +40,7 @@ def main():
         cfg = SuiConfig.from_config_file(cfg_file[0])
     else:
         cfg = SuiConfig.default()
-
+    print(parsed)
     try:
         parsed.excludes = set(parsed.excludes) if parsed.excludes else set()
         parsed.includes = set(parsed.includes) if parsed.includes else set()
@@ -49,8 +49,10 @@ def main():
         )
         package_gen = PackageGen(
             package_ir=ir_builder.generate_ir(),
-            target_path=Path(os.path.expanduser("~/")),
-            template_path=_MODEL_PATH,
+            target_path=Path(parsed.root_path),
+            template_path=_TEMPLATE_PATH,
+            overwrite_modules=parsed.overwrite_modules,
+            use_async=parsed.use_async,
         )
         package_gen.generate()