+++ pysui_gadgets/dsl/generators/package_gen.py
@@ -13,6 +13,8 @@
 
 """pysui_gadgets: DSL Struct to Python class generator."""
 
+import os
+import shutil
 import ast
 import copy
 from pathlib import Path
@@ -31,15 +33,19 @@ class StructureGen:
         """."""
         self.struct_irs = structure_irs
 
-    def _get_method_for(self, for_method: str, ast_node: ast.ClassDef) -> Union[ast.FunctionDef, Exception]:
+    def _get_method_for(
+        self, for_method: str, ast_node: ast.ClassDef
+    ) -> Union[ast.FunctionDef, ast.AsyncFunctionDef, Exception]:
         """."""
         for node in ast_node.body:
-            if isinstance(node, ast.FunctionDef):
+            if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)):
                 if node.name == for_method:
                     return node
         raise AttributeError(f"No method found for {for_method}")
 
-    def update_init_node(self, struct_ir: StructIR, node: ast.FunctionDef) -> ast.FunctionDef:
+    def update_init_node(
+        self, struct_ir: StructIR, node: Union[ast.FunctionDef, ast.AsyncFunctionDef]
+    ) -> Union[ast.FunctionDef, ast.AsyncFunctionDef]:
         """_update_init Inserts __init__ parameters and property assignments.
 
         :param node: The AST __init__function definition
@@ -51,7 +57,9 @@ class StructureGen:
         node.body.extend([ast.parse(f"self.{x.name}:{x.type_signature} = {x.name}").body[0] for x in struct_ir.fields])
         return node
 
-    def update_instance_node(self, struct_ir: StructIR, node: ast.FunctionDef) -> ast.FunctionDef:
+    def update_instance_node(
+        self, struct_ir: StructIR, node: Union[ast.FunctionDef, ast.AsyncFunctionDef]
+    ) -> Union[ast.FunctionDef, ast.AsyncFunctionDef]:
         """_update_loader Sets new struct class return type in classmethod.
 
         :param node: The AST get_instance classmethod function definition.
@@ -60,7 +68,9 @@ class StructureGen:
         node.returns = ast.Name(id=f'"{struct_ir.name}"')
         return node
 
-    def update_property_node(self, struct_ir: StructIR, node: ast.FunctionDef) -> list[ast.FunctionDef]:
+    def update_property_node(
+        self, struct_ir: StructIR, node: Union[ast.FunctionDef, ast.AsyncFunctionDef]
+    ) -> list[Union[ast.FunctionDef, ast.AsyncFunctionDef]]:
         """."""
         new_props = []
         for field in struct_ir.fields:
@@ -68,7 +78,7 @@ class StructureGen:
                 as_arg_node = copy.deepcopy(node)
                 as_arg_node.name = f"{field.name}_argument"
                 if field.arg_converter_returns:
-                    as_arg_node.returns = ast.parse(field.arg_converter_returns).body[0].value
+                    as_arg_node.returns = ast.Name(id=field.arg_converter_returns)
                 if field.as_arg_converter:
                     as_arg_node.body[1] = ast.parse(f"return {field.as_arg_converter}")
                 new_props.append(as_arg_node)
@@ -76,7 +86,7 @@ class StructureGen:
                 as_type_node = copy.deepcopy(node)
                 as_type_node.name = f"{field.name}_suitype"
                 if field.type_converter_returns:
-                    as_type_node.returns = ast.parse(field.type_converter_returns).body[0].value
+                    as_type_node.returns = ast.Name(id=field.type_converter_returns)
                 if field.as_type_converter:
                     as_type_node.body[1] = ast.parse(f"return {field.as_type_converter}")
                 new_props.append(as_type_node)
@@ -196,16 +206,100 @@ class ModuleGen:
 
 
 class PackageGen:
-    """PacageGen creates the package layout."""
+    """PacageGen creates the package layout and content."""
 
-    def __init__(self, *, package_ir: PackageIR, target_path: Path, template_path: Path):
-        """."""
+    def __init__(
+        self, *, package_ir: PackageIR, target_path: Path, template_path: Path, overwrite_modules: bool, use_async: bool
+    ):
+        """__init__ Generates structure and module entry point method files.
+
+        :param package_ir: The package/module internal representation from pysui_gadgets.dsl.ir.modir
+        :type package_ir: PackageIR
+        :param target_path: Path to where package/module(s) will be written to
+        :type target_path: Path
+        :param template_path: Path to internal templates for structs/modules
+        :type template_path: Path
+        :param overwrite_modules: If True, existing path/module(s) can be overwritten
+        :type overwrite_modules: bool
+        :param use_async: If True, an ``asynchronous`` flavor of the template(s) are used
+        :type use_async: bool
+        :raises ValueError: If something goes awry
+        """
         self.package_ir = package_ir
         self.target_path = target_path.joinpath(package_ir.name)
-        self.template_path = template_path
+        # synch vs asynch
+        if use_async:
+            self.template_path = template_path.joinpath("async_model.py")
+        else:
+            self.template_path = template_path.joinpath("sync_model.py")
+        self.overwrite_modules = overwrite_modules
+        # Get the reusable __init__.py for package and modules
+        with open(template_path.joinpath("__init__.py"), "r", encoding="utf-8") as init_file:
+            self.init_content = init_file.read()
+        # Upfront testing of target state
+        self.package_exists, self.modules_exist = self._check_target()
+        if self.modules_exist and not self.overwrite_modules:
+            raise ValueError(
+                f"Modules exist for package {package_ir.name} and overwrite option = {self.overwrite_modules}"
+            )
+
+    def _check_target(self) -> tuple[bool, bool]:
+        """_check_target Checks if package and modules already exist.
+
+        :return: True or False for packcage folder and module folder respectivly
+        :rtype: tuple[bool, bool]
+        """
+        have_package = self.target_path.exists()
+        if have_package:
+            mod_names = [x.name for x in self.package_ir.modules]
+            filtered = list(filter(lambda x: self.target_path.joinpath(x).exists(), mod_names))
+            return (have_package, len(filtered) > 0)
+        return (have_package, False)
+
+    def _write_init_py(self, base_path: Path) -> Union[None, Exception]:
+        """_write_init_py Writes an ``__init__.py`` at base_path.
+
+        :param base_path: Path to where ``__init__.py`` is written
+        :type base_path: Path
+        :return: Nothing or exception from file open/write
+        :rtype: Union[None, Exception]
+        """
+        with open(base_path.joinpath("__init__.py"), "w", encoding="utf-8") as init_file:
+            init_file.write(self.init_content)
+
+    def _create_python_package(self) -> Union[None, Exception]:
+        """_create_python_package Creates the initial package if not already exist.
+
+        :return: Nothing or exception from path create or file open/write
+        :rtype: Union[None, Exception]
+        """
+        if not self.package_exists:
+            print(f"Creating package {self.package_ir.name}")
+            os.mkdir(self.target_path)
+            self._write_init_py(self.target_path)
+
+    def _create_python_module(self, module_name: str, ast_node: ast.Module) -> Union[None, Exception]:
+        """_create_python_module Creates a module package folder and content.
+
+        :param module_name: Name of module (from move module)
+        :type module_name: str
+        :param ast_node: The python AST containing recently generated module content
+        :type ast_node: ast.Module
+        :return: Nothing or exception from path create or file open/writes
+        :rtype: Union[None, Exception]
+        """
+        base_path = self.target_path.joinpath(module_name)
+        base_module = base_path.joinpath(module_name + ".py")
+        if base_path.exists() and self.overwrite_modules:
+            shutil.rmtree(base_path)
+        os.mkdir(base_path)
+        with open(base_module, "w", encoding="utf-8") as module_file:
+            module_file.write(ast.unparse(ast_node))
+        self._write_init_py(base_path)
 
     def generate(self):
-        """."""
+        """generate Generate python AST encapsulating move package/functions and struct details."""
+        self._create_python_package()
         for module in self.package_ir.modules:
             mod_gen = ModuleGen(
                 package_id=self.package_ir.name[1:],
@@ -213,5 +307,5 @@ class PackageGen:
                 package_path=self.target_path,
                 template_path=self.template_path,
             )
-            mod_ast = mod_gen.generate()
-            print(ast.unparse(mod_ast))
+            self._create_python_module(module.name, mod_gen.generate())
+            # print(ast.unparse(mod_gen.generate()))