+++ CHANGELOG.md
@@ -4,7 +4,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unpublished]
+
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+## [0.0.3] - 2022-12-15
+
+### Added
+- [enhancement](https://github.com/FrankC01/pysui_gadgets/issues/1)
+- [enhancement](https://github.com/FrankC01/pysui_gadgets/issues/2)
+
+### Fixed
+- ObjectRead transition in `templates/sync_model.py`
+
+### Changed
+- Removed cruft from template
+- Reducedd calls to `ast.parse`
+
+### Removed
+
+## [0.0.2] - 2022-12-15
 
 ### Added
 - package: Commands to inspect SUI package meta-data